initial round of fixes

+++ include/core/buffer.hpp
@@ -28,13 +28,13 @@ namespace Vulkan::Core {
         /// @throws std::invalid_argument if the device or buffer size is invalid
         /// @throws ls::vulkan_error if object creation fails.
         ///
-        Buffer(const Device& device, uint32_t size, std::vector<uint8_t> data,
+        Buffer(const Device& device, size_t size, std::vector<uint8_t> data,
             VkBufferUsageFlags usage);
 
         /// Get the Vulkan handle.
         [[nodiscard]] auto handle() const { return *this->buffer; }
         /// Get the size of the buffer.
-        [[nodiscard]] uint32_t getSize() const { return this->size; }
+        [[nodiscard]] size_t getSize() const { return this->size; }
 
         /// Check whether the object is valid.
         [[nodiscard]] bool isValid() const { return static_cast<bool>(this->buffer); }
@@ -51,7 +51,7 @@ namespace Vulkan::Core {
         std::shared_ptr<VkBuffer> buffer;
         std::shared_ptr<VkDeviceMemory> memory;
 
-        uint32_t size;
+        size_t size;
     };
 
 }
